+++ src/main/java/ar/com/draimo/jitws/model/ViajeTerceroCombustible.java
@@ -65,7 +65,7 @@ public class ViajeTerceroCombustible extends ObjetoGenerico {
     
     //Define la cantidad
     @Column(name = "cantidad", nullable = false)
-    private short cantidad;
+    private BigDecimal cantidad;
     
     //Define el precio unitario
     @Column(name = "precioUnitario", nullable = false)
@@ -149,11 +149,11 @@ public class ViajeTerceroCombustible extends ObjetoGenerico {
         this.insumo = insumo;
     }
 
-    public short getCantidad() {
+    public BigDecimal getCantidad() {
         return cantidad;
     }
 
-    public void setCantidad(short cantidad) {
+    public void setCantidad(BigDecimal cantidad) {
         this.cantidad = cantidad;
     }
